Get directory name before going gc-free.  Throw error if directory fails to open.  Remove some compile warnings

+++ src/hx/CppiaBuiltin.cpp
@@ -174,7 +174,7 @@ struct ArrayBuiltin : public ArrayBuiltinBase
 
          case afPop:
          //case afUnshift:
-            if (ExprTypeOf<ELEM>::value==etString)
+            if (ExprTypeOf<ELEM>::value==(int)etString)
                return etString;
             return etObject;
